refactor: update recipeBook ModelManager and recipeUserPref
